% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{add_viscometer}
\alias{add_viscometer}
\title{Add a viscometer record to the local registry.}
\usage{
add_viscometer(viscometer)
}
\arguments{
\item{viscometer}{A viscometer record tibble.}
}
\value{
The inserted viscometer record.
}
\description{
Add a viscometer record to the local registry.
}
