% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{archive_viscometer}
\alias{archive_viscometer}
\title{Archive a viscometer in the local registry.}
\usage{
archive_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
The updated viscometer record.
}
\description{
Archive a viscometer in the local registry.
}
