% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_viscometer_id}
\alias{format_viscometer_id}
\title{Format a viscometer ID}
\usage{
format_viscometer_id(viscometer_size, serial_number, fn = NULL)
}
\arguments{
\item{viscometer_size}{Integer viscometer size number (1-999), per the
ASTM viscometer size designation used to construct the ID prefix.}

\item{serial_number}{Character serial number, used as the ID suffix.}

\item{fn}{Character scalar naming the calling function, used in error
messages. Defaults to the name of the calling function.}
}
\value{
A character scalar viscometer ID in \code{NNN-NNNNN} format.
}
\description{
Builds the canonical \code{NNN-NNNNN} viscometer ID (three-digit size prefix,
hyphen, serial number) used as the primary key throughout the registry.
}
\examples{
format_viscometer_id(200, "12345")

}
