% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{get_viscometer}
\alias{get_viscometer}
\title{Get one viscometer by identifier.}
\usage{
get_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
A one-row tibble for the requested viscometer.
}
\description{
Get one viscometer by identifier.
}
