% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{list_active_viscometers}
\alias{list_active_viscometers}
\title{List active viscometers in the local registry.}
\usage{
list_active_viscometers()
}
\value{
A tibble of non-archived viscometer records.
}
\description{
List active viscometers in the local registry.
}
