% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{list_archived_viscometers}
\alias{list_archived_viscometers}
\title{List archived viscometers in the local registry.}
\usage{
list_archived_viscometers()
}
\value{
A tibble of archived viscometer records.
}
\description{
List archived viscometers in the local registry.
}
\examples{
old_opt <- options(kinvicalc.reference_db_path = tempfile(fileext = ".db"))

list_archived_viscometers()

options(old_opt)
}
