% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing infix operator.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value to test.}

\item{y}{Fallback value returned when \code{x} is \code{NULL}.}
}
\value{
\code{y} if \code{x} is \code{NULL}, otherwise \code{x}.
}
\description{
Null-coalescing infix operator.
}
\examples{
kinvicalc:::`\%||\%`(NULL, "default")
kinvicalc:::`\%||\%`("value", "default")
}
\keyword{internal}
