% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-core.R
\name{reference_db_connection}
\alias{reference_db_connection}
\title{Get the local SQLite reference database connection.}
\usage{
reference_db_connection()
}
\value{
An open DBI connection.
}
\description{
Get the local SQLite reference database connection.
}
\keyword{internal}
