% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{remove_viscometer}
\alias{remove_viscometer}
\title{Remove a viscometer from the local registry.}
\usage{
remove_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
Invisibly TRUE when removed.
}
\description{
Remove a viscometer from the local registry.
}
