% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{reset_viscometer_use_count}
\alias{reset_viscometer_use_count}
\title{Reset viscometer use count since deep cleaning.}
\usage{
reset_viscometer_use_count(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
A one-row tibble with the updated use counters for the viscometer.
}
\description{
Resets the post-cleaning usage counter for a viscometer and records the
current time as the most recent deep-clean timestamp.
}
\examples{
\dontrun{
reset_viscometer_use_count("007-00007")
}
}
