% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the kinvicalc Shiny application.}
\usage{
run_app()
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
Run the kinvicalc Shiny application.
}
\examples{
if (interactive()) {
  run_app()
}
}
