% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-core.R
\name{save_reference_data}
\alias{save_reference_data}
\title{Save reference data to the local database.}
\usage{
save_reference_data(viscometers = NULL, sample_types = NULL)
}
\arguments{
\item{viscometers}{A tibble of viscometer records.}

\item{sample_types}{A tibble of sample-type rules.}
}
\value{
Invisibly TRUE when saved.
}
\description{
Save reference data to the local database.
}
