% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{session_results_table}
\alias{session_results_table}
\title{Create a session results table.}
\usage{
session_results_table(results)
}
\arguments{
\item{results}{A list of result objects.}
}
\value{
A tibble summarizing the results.
}
\description{
Create a session results table.
}
