% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-types.R
\name{add_sample_type_rule}
\alias{add_sample_type_rule}
\title{Add or replace a sample type precision rule.}
\usage{
add_sample_type_rule(rule)
}
\arguments{
\item{rule}{A sample type rule record (see \code{validate_sample_type_rule()}
for required columns).}
}
\value{
The inserted rule.
}
\description{
Add or replace a sample type precision rule.
}
