% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{format_sample_type_label_app}
\alias{format_sample_type_label_app}
\title{Run the kinvicalc Shiny application.}
\usage{
format_sample_type_label_app(sample_type)
}
\arguments{
\item{sample_type}{A sample type code.}
}
\description{
Run the kinvicalc Shiny application.
}
