% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{get_viscometer}
\alias{get_viscometer}
\title{Get one viscometer by identifier.}
\usage{
get_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
A one-row tibble for the requested viscometer.
}
\description{
Get one viscometer by identifier.
}
\examples{
old_opt <- options(kinvicalc.reference_db_path = tempfile(fileext = ".db"))

add_viscometer(
  build_viscometer_record(
    viscometer_size = 1,
    serial_number = "00001",
    factor_40_top = 0.025,
    factor_40_bottom = 0.029,
    factor_100_top = 0.016,
    factor_100_bottom = 0.018
  )
)
get_viscometer("001-00001")

options(old_opt)
}
