% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{list_viscometers}
\alias{list_viscometers}
\title{List all viscometers in the local registry.}
\usage{
list_viscometers()
}
\value{
A tibble of viscometer records.
}
\description{
List all viscometers in the local registry.
}
