% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-core.R
\name{load_reference_data}
\alias{load_reference_data}
\title{Load reference data from the local database.}
\usage{
load_reference_data()
}
\value{
A list containing viscometers and sample_types.
}
\description{
Load reference data from the local database.
}
