% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{lock_result}
\alias{lock_result}
\title{Lock a result in the session state.}
\usage{
lock_result(result)
}
\arguments{
\item{result}{A result object.}
}
\value{
The same result object with a locked flag.
}
\description{
Lock a result in the session state.
}
