% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-core.R
\name{MIN_FLOW_TIME_S}
\alias{MIN_FLOW_TIME_S}
\title{Minimum flow time, in seconds, below which ASTM D445-26 6.1.2/D446-24 8.2.1
requires a kinetic energy correction to be applied.}
\usage{
MIN_FLOW_TIME_S
}
\description{
Minimum flow time, in seconds, below which ASTM D445-26 6.1.2/D446-24 8.2.1
requires a kinetic energy correction to be applied.
}
\keyword{internal}
