% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{remove_viscometer}
\alias{remove_viscometer}
\title{Remove a viscometer from the local registry.}
\usage{
remove_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
Invisibly TRUE when removed.
}
\description{
Remove a viscometer from the local registry.
}
\examples{
old_opt <- options(kinvicalc.reference_db_path = tempfile(fileext = ".db"))

add_viscometer(
  build_viscometer_record(
    viscometer_size = 1,
    serial_number = "00001",
    factor_40_top = 0.025,
    factor_40_bottom = 0.029,
    factor_100_top = 0.016,
    factor_100_bottom = 0.018
  )
)
remove_viscometer("001-00001")

options(old_opt)
}
