% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{session_results_table}
\alias{session_results_table}
\title{Create a session results table.}
\usage{
session_results_table(results)
}
\arguments{
\item{results}{A list of result objects.}
}
\value{
A tibble summarizing the results.
}
\description{
Create a session results table.
}
\examples{
old_opt <- options(kinvicalc.reference_db_path = tempfile(fileext = ".db"))

add_viscometer(
  build_viscometer_record(
    viscometer_size = 1,
    serial_number = "00001",
    factor_40_top = 0.025,
    factor_40_bottom = 0.029,
    factor_100_top = 0.016,
    factor_100_bottom = 0.018
  )
)
result <- build_sample_result(
  viscometer_id = "001-00001",
  sample_type = "unlisted",
  analysis_temperature_c = 40,
  time_1 = 232,
  time_2 = 200
)

session_results_table(list(result))

options(old_opt)
}
