% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{unarchive_viscometer}
\alias{unarchive_viscometer}
\title{Unarchive a viscometer in the local registry.}
\usage{
unarchive_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{A viscometer identifier.}
}
\value{
The updated viscometer record.
}
\description{
Unarchive a viscometer in the local registry.
}
