% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscometer.R
\name{validate_viscometer}
\alias{validate_viscometer}
\title{Validate a viscometer identifier.}
\usage{
validate_viscometer(viscometer_id)
}
\arguments{
\item{viscometer_id}{Viscometer ID string.}
}
\value{
TRUE if valid, otherwise FALSE.
}
\description{
Validate a viscometer identifier.
}
\examples{
validate_viscometer("200-12345")
validate_viscometer("not-an-id")
}
