% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\name{validate_viscometer_record}
\alias{validate_viscometer_record}
\title{Validate a viscometer record.}
\usage{
validate_viscometer_record(viscometer)
}
\arguments{
\item{viscometer}{A data frame-like object representing one viscometer.}
}
\value{
The validated viscometer tibble.
}
\description{
Checks that a viscometer record contains all required columns and valid values.
}
\examples{
validate_viscometer_record(
  tibble::tibble(
    viscometer_size = 75,
    serial_number = "294",
    status = "active",
    factor_40_top = 0.00842,
    factor_40_bottom = 0.00609,
    factor_100_top = 0.00847,
    factor_100_bottom = 0.00614
  )
)
}
